refactor: classify the file for syncing the scroll
